.db coded

+++ src/main/java/com/bit/backend/repositories/CustomerRepository.java
@@ -3,6 +3,6 @@ package com.bit.backend.repositories;
 import com.bit.backend.entities.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<CustomerEntity, Integer> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity, Long> {
 
 }
